inquiry flag fixes
